🐛 fix : 회원가입 시 멤버 정보 덮어쓰는 오류 수정 Member id 생성 방식을 Auto-Increment에서 현재 nano초로 변경

+++ backend/src/main/java/org/a204/hourgoods/global/security/jwt/JwtAuthenticationProvider.java
@@ -50,7 +50,7 @@ public class JwtAuthenticationProvider implements AuthenticationProvider {
 		if (principal != null) {
 			userDetails = new MemberDetails(Member.builder()
 				.email((String)claims.get("email"))
-				.id(Long.valueOf((Integer)claims.get("id")))
+				.id(Long.valueOf((Long)claims.get("id")))
 				.build());
 		} else {
 			userDetails = new AdminDetails((String)claims.get("adminId"));
